1. Navigation UI drawer is added. 2. Each activity is mapped with navigation drawer items. (Added java code) 3. Customized toolbar for each activity is added. 4. Theme color changed. 5. More items added in navigation item list.
