+++ app/src/main/java/com/BugBazaar/ui/Contact_us.java
@@ -3,6 +3,8 @@ package com.BugBazaar.ui;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+import android.view.View;
+import android.widget.TextView;
 
 import com.BugBazaar.R;
 
@@ -12,5 +14,13 @@ public class Contact_us extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_contact_us);
+        //Toolbar title set
+        TextView toolbarTitle = findViewById(R.id.toolbarTitle);
+        toolbarTitle.setText("Contact-Us");
+    }
+
+    //Code to handle backbutton
+    public void onBackButtonClick(View view) {
+        onBackPressed(); // Navigate back to the previous activity
     }
 }
