+++ app/src/main/java/com/BugBazaar/ui/TermsAndConditionsActivity.java
@@ -3,6 +3,7 @@ package com.BugBazaar.ui;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+import android.view.View;
 import android.webkit.WebChromeClient;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
@@ -10,6 +11,7 @@ import android.webkit.WebViewClient;
 import com.BugBazaar.R;
 import com.BugBazaar.utils.AppConstants;
 import android.webkit.SafeBrowsingResponse;
+import android.widget.TextView;
 
 public class TermsAndConditionsActivity extends AppCompatActivity {
     WebView webView;
@@ -21,6 +23,10 @@ public class TermsAndConditionsActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_terms_and_conditions);
         webView = findViewById(R.id.terms_condition_view);
+
+        //Toolbar title set
+        TextView toolbarTitle = findViewById(R.id.toolbarTitle);
+        toolbarTitle.setText("Terms & Conditions");
 
         setupwebview(webView);
         if (getIntent().getExtras() != null) {
@@ -67,5 +73,9 @@ public class TermsAndConditionsActivity extends AppCompatActivity {
         }
 
 
+    }
+    //Code to handle backbutton
+    public void onBackButtonClick(View view) {
+        onBackPressed(); // Navigate back to the previous activity
     }
 }
