+++ app/src/main/res/layout/activity_main.xml
@@ -12,7 +12,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="BugBazaar"
-        android:textColor="#FFFFFF"
         android:textSize="40sp"
         android:padding="20dp"
         android:layout_gravity="center">
@@ -48,7 +47,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Login"
-        android:layout_gravity="center"
-        android:backgroundTint="@color/teal_200" />
+        android:layout_gravity="center" />
 
 </LinearLayout>
