+++ app/src/main/res/layout/activity_navigation_drawer_dashboard.xml
@@ -13,19 +13,20 @@
         android:id="@+id/main_content"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
+        <!--Toolbar-->
+        <include layout="@layout/nav_toolbar_main"></include>
 
-        <androidx.appcompat.widget.Toolbar
-            android:id="@+id/toolbar"
+        <!--MainContent-->
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/fragment_container"
             android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:background="?attr/colorPrimary" />
-
+            android:layout_height="match_parent" />
         <TextView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:layout_gravity="center"
             android:gravity="center"
-            android:text="XDXDXDXD"
+            android:text="DashBoard"
             android:textSize="34sp" />
     </FrameLayout>
 
@@ -36,6 +37,5 @@
         android:layout_gravity="start"
         app:headerLayout="@layout/nav_header_layout"
         app:menu="@menu/navigation_items" />
-
 
 </androidx.drawerlayout.widget.DrawerLayout>
