+++ app/src/main/res/layout/activity_terms_and_conditions.xml
@@ -4,12 +4,33 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:gravity="top"
+    android:orientation="vertical"
     tools:context=".ui.TermsAndConditionsActivity">
 
 
-    <WebView
+    <include
+        android:id="@+id/include"
+        layout="@layout/nav_toolbar_sub"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/terms_condition_view" />
+        android:layout_height="wrap_content"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.0"></include>
+
+    <WebView
+        android:id="@+id/terms_condition_view"
+        android:layout_width="match_parent"
+        android:layout_height="675dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/include"
+        app:layout_constraintVertical_bias="0.0">
+
+    </WebView>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
