+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,6 @@
     <string name="app_name">BugBazaar</string>
     <string name="navigation_drawer_open">Open Drawer</string>
     <string name="navigation_drawer_close">Close Drawer</string>
+    <string name="my_profile">My Profile</string>
+
 </resources>
