+++ app/src/main/res/values/themes.xml
@@ -2,15 +2,14 @@
     <!-- Base application theme. -->
     <style name="Theme.BugBazaar" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/teal_700</item>
-        <item name="colorPrimaryVariant">#00BCD4</item>
-        <item name="colorOnPrimary">@color/white</item>
+        <item name="colorPrimary">#3a5e7d</item>
+        <item name="colorPrimaryVariant">#3a5e7d</item>
+        <item name="colorOnPrimary">#FBEEC1</item>
         <!-- Secondary brand color. -->
         <item name="colorSecondary">@color/teal_200</item>
         <item name="colorSecondaryVariant">@color/teal_700</item>
         <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
-        <item name="android:statusBarColor">@color/teal_700</item>
-        <!-- Customize your theme here. -->
+        <item name="android:statusBarColor">#3a5e7d</item>
     </style>
 </resources>
